move functionality to dense
- diagonal extractable
- scaled identity addable
- add/sub scaled diagonal
