+++ common/unified/matrix/dense_kernels.cpp
@@ -85,6 +85,81 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
     GKO_DECLARE_DENSE_COUNT_NONZEROS_PER_ROW_KERNEL_SIZE_T);
 
 
+template <typename ValueType>
+void extract_diagonal(std::shared_ptr<const DefaultExecutor> exec,
+                      matrix::view::dense<const ValueType> orig,
+                      matrix::Diagonal<ValueType>* diag)
+{
+    run_kernel(
+        exec,
+        [] GKO_KERNEL(auto i, auto orig, auto diag) { diag[i] = orig(i, i); },
+        diag->get_size()[0], orig, diag->get_values());
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_EXTRACT_DIAGONAL_KERNEL);
+
+
+template <typename ValueType>
+void add_scaled_diag(std::shared_ptr<const DefaultExecutor> exec,
+                     matrix::view::dense<const ValueType> alpha,
+                     const matrix::Diagonal<ValueType>* x,
+                     matrix::view::dense<ValueType> y)
+{
+    const auto diag_values = x->get_const_values();
+    run_kernel(
+        exec,
+        [] GKO_KERNEL(auto i, auto alpha, auto diag, auto y) {
+            if (is_nonzero(alpha[0])) {
+                y(i, i) += alpha[0] * diag[i];
+            }
+        },
+        x->get_size()[0], alpha.values, x->get_const_values(), y);
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_ADD_SCALED_DIAG_KERNEL);
+
+
+template <typename ValueType>
+void sub_scaled_diag(std::shared_ptr<const DefaultExecutor> exec,
+                     matrix::view::dense<const ValueType> alpha,
+                     const matrix::Diagonal<ValueType>* x,
+                     matrix::view::dense<ValueType> y)
+{
+    const auto diag_values = x->get_const_values();
+    run_kernel(
+        exec,
+        [] GKO_KERNEL(auto i, auto alpha, auto diag, auto y) {
+            if (is_nonzero(alpha[0])) {
+                y(i, i) -= alpha[0] * diag[i];
+            }
+        },
+        x->get_size()[0], alpha.values, x->get_const_values(), y);
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_SUB_SCALED_DIAG_KERNEL);
+
+
+template <typename ValueType, typename ScalarType>
+void add_scaled_identity(std::shared_ptr<const DefaultExecutor> exec,
+                         matrix::view::dense<const ScalarType> alpha,
+                         matrix::view::dense<const ScalarType> beta,
+                         matrix::view::dense<ValueType> mtx)
+{
+    run_kernel(
+        exec,
+        [] GKO_KERNEL(auto row, auto col, auto alpha, auto beta, auto mtx) {
+            mtx(row, col) = beta[0] * mtx(row, col);
+            if (row == col) {
+                mtx(row, row) += alpha[0];
+            }
+        },
+        mtx.size, alpha.values, beta.values, mtx);
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_SCALAR_TYPE(
+    GKO_DECLARE_DENSE_ADD_SCALED_IDENTITY_KERNEL);
+
+
 }  // namespace dense
 }  // namespace GKO_DEVICE_NAMESPACE
 }  // namespace kernels
