+++ core/device_hooks/common_kernels.inc.cpp
@@ -471,6 +471,10 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
     GKO_DECLARE_DENSE_COUNT_NONZEROS_PER_ROW_KERNEL_SIZE_T);
 GKO_STUB_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_DENSE_COUNT_NONZERO_BLOCKS_PER_ROW_KERNEL);
+GKO_STUB_VALUE_AND_SCALAR_TYPE(GKO_DECLARE_DENSE_ADD_SCALED_IDENTITY_KERNEL);
+GKO_STUB_VALUE_TYPE(GKO_DECLARE_DENSE_ADD_SCALED_DIAG_KERNEL);
+GKO_STUB_VALUE_TYPE(GKO_DECLARE_DENSE_SUB_SCALED_DIAG_KERNEL);
+GKO_STUB_VALUE_TYPE(GKO_DECLARE_DENSE_EXTRACT_DIAGONAL_KERNEL);
 
 
 }  // namespace dense
@@ -485,10 +489,6 @@ GKO_STUB_VALUE_AND_SCALAR_TYPE(GKO_DECLARE_MULTIVECTOR_SCALE_KERNEL);
 GKO_STUB_VALUE_AND_SCALAR_TYPE(GKO_DECLARE_MULTIVECTOR_INV_SCALE_KERNEL);
 GKO_STUB_VALUE_AND_SCALAR_TYPE(GKO_DECLARE_MULTIVECTOR_ADD_SCALED_KERNEL);
 GKO_STUB_VALUE_AND_SCALAR_TYPE(GKO_DECLARE_MULTIVECTOR_SUB_SCALED_KERNEL);
-GKO_STUB_VALUE_AND_SCALAR_TYPE(
-    GKO_DECLARE_MULTIVECTOR_ADD_SCALED_IDENTITY_KERNEL);
-GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_ADD_SCALED_DIAG_KERNEL);
-GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_SUB_SCALED_DIAG_KERNEL);
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_COMPUTE_DOT_KERNEL);
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_COMPUTE_DOT_DISPATCH_KERNEL);
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_COMPUTE_CONJ_DOT_KERNEL);
@@ -529,7 +529,6 @@ GKO_STUB_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_MULTIVECTOR_NONSYMM_SCALE_PERMUTE_KERNEL);
 GKO_STUB_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_MULTIVECTOR_INV_NONSYMM_SCALE_PERMUTE_KERNEL);
-GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_EXTRACT_DIAGONAL_KERNEL);
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_INPLACE_ABSOLUTE_DENSE_KERNEL);
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_OUTPLACE_ABSOLUTE_DENSE_KERNEL);
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_MAKE_COMPLEX_KERNEL);
