+++ core/matrix/multivector.cpp
@@ -38,8 +38,6 @@ GKO_REGISTER_OPERATION(scale, multivector::scale);
 GKO_REGISTER_OPERATION(inv_scale, multivector::inv_scale);
 GKO_REGISTER_OPERATION(add_scaled, multivector::add_scaled);
 GKO_REGISTER_OPERATION(sub_scaled, multivector::sub_scaled);
-GKO_REGISTER_OPERATION(add_scaled_diag, multivector::add_scaled_diag);
-GKO_REGISTER_OPERATION(sub_scaled_diag, multivector::sub_scaled_diag);
 GKO_REGISTER_OPERATION(compute_dot, multivector::compute_dot_dispatch);
 GKO_REGISTER_OPERATION(compute_conj_dot,
                        multivector::compute_conj_dot_dispatch);
@@ -74,7 +72,6 @@ GKO_REGISTER_OPERATION(inv_row_scale_permute,
 GKO_REGISTER_OPERATION(inv_col_scale_permute,
                        multivector::inv_col_scale_permute);
 GKO_REGISTER_OPERATION(fill_in_matrix_data, multivector::fill_in_matrix_data);
-GKO_REGISTER_OPERATION(extract_diagonal, multivector::extract_diagonal);
 GKO_REGISTER_OPERATION(inplace_absolute_dense,
                        multivector::inplace_absolute_dense);
 GKO_REGISTER_OPERATION(outplace_absolute_dense,
@@ -82,7 +79,6 @@ GKO_REGISTER_OPERATION(outplace_absolute_dense,
 GKO_REGISTER_OPERATION(make_complex, multivector::make_complex);
 GKO_REGISTER_OPERATION(get_real, multivector::get_real);
 GKO_REGISTER_OPERATION(get_imag, multivector::get_imag);
-GKO_REGISTER_OPERATION(add_scaled_identity, multivector::add_scaled_identity);
 
 
 }  // anonymous namespace
@@ -124,33 +120,14 @@ void MultiVector<ValueType>::validate_data() const
 
 
 template <typename ValueType>
-void MultiVector<ValueType>::apply_impl(const LinOp* b, LinOp* x) const
-{
-    precision_dispatch_real_complex<ValueType>(
-        [this](auto dense_b, auto dense_x) {
-            this->get_executor()->run(multivector::make_simple_apply(
-                this->get_const_device_view(), dense_b->get_const_device_view(),
-                dense_x->get_device_view()));
-        },
-        b, x);
-}
+void MultiVector<ValueType>::apply_impl(const LinOp* b,
+                                        LinOp* x) const GKO_NOT_IMPLEMENTED;
 
 
 template <typename ValueType>
 void MultiVector<ValueType>::apply_impl(const LinOp* alpha, const LinOp* b,
-                                        const LinOp* beta, LinOp* x) const
-{
-    precision_dispatch_real_complex<ValueType>(
-        [this](auto dense_alpha, auto dense_b, auto dense_beta, auto dense_x) {
-            this->get_executor()->run(multivector::make_apply(
-                dense_alpha->get_const_device_view(),
-                this->get_const_device_view(), dense_b->get_const_device_view(),
-                dense_beta->get_const_device_view(),
-                dense_x->get_device_view()));
-        },
-        alpha, b, beta, x);
-}
-
+                                        const LinOp* beta,
+                                        LinOp* x) const GKO_NOT_IMPLEMENTED;
 
 template <typename ValueType>
 void MultiVector<ValueType>::fill(const ValueType value)
@@ -322,20 +299,11 @@ void MultiVector<ValueType>::add_scaled_impl(const LinOp* alpha, const LinOp* b)
                 ->get_const_device_view(),
             dynamic_cast<complex_type*>(this)->get_device_view()));
     } else {
-        if (dynamic_cast<const Diagonal<ValueType>*>(b)) {
-            exec->run(multivector::make_add_scaled_diag(
-                make_temporary_conversion<ValueType>(alpha)
-                    ->get_const_device_view(),
-                dynamic_cast<const Diagonal<ValueType>*>(b),
-                this->get_device_view()));
-        } else {
-            exec->run(multivector::make_add_scaled(
-                make_temporary_conversion<ValueType>(alpha)
-                    ->get_const_device_view(),
-                make_temporary_conversion<ValueType>(b)
-                    ->get_const_device_view(),
-                this->get_device_view()));
-        }
+        exec->run(multivector::make_add_scaled(
+            make_temporary_conversion<ValueType>(alpha)
+                ->get_const_device_view(),
+            make_temporary_conversion<ValueType>(b)->get_const_device_view(),
+            this->get_device_view()));
     }
 }
 
@@ -360,20 +328,11 @@ void MultiVector<ValueType>::sub_scaled_impl(const LinOp* alpha, const LinOp* b)
                 ->get_const_device_view(),
             dynamic_cast<complex_type*>(this)->get_device_view()));
     } else {
-        if (dynamic_cast<const Diagonal<ValueType>*>(b)) {
-            exec->run(multivector::make_sub_scaled_diag(
-                make_temporary_conversion<ValueType>(alpha)
-                    ->get_const_device_view(),
-                dynamic_cast<const Diagonal<ValueType>*>(b),
-                this->get_device_view()));
-        } else {
-            exec->run(multivector::make_sub_scaled(
-                make_temporary_conversion<ValueType>(alpha)
-                    ->get_const_device_view(),
-                make_temporary_conversion<ValueType>(b)
-                    ->get_const_device_view(),
-                this->get_device_view()));
-        }
+        exec->run(multivector::make_sub_scaled(
+            make_temporary_conversion<ValueType>(alpha)
+                ->get_const_device_view(),
+            make_temporary_conversion<ValueType>(b)->get_const_device_view(),
+            this->get_device_view()));
     }
 }
 
@@ -1612,31 +1571,6 @@ void MultiVector<ValueType>::scale_permute(
 
 
 template <typename ValueType>
-void MultiVector<ValueType>::extract_diagonal(
-    ptr_param<Diagonal<ValueType>> output) const
-{
-    auto exec = this->get_executor();
-    const auto diag_size = std::min(this->get_size()[0], this->get_size()[1]);
-    GKO_ASSERT_EQ(output->get_size()[0], diag_size);
-
-    exec->run(multivector::make_extract_diagonal(
-        this->get_const_device_view(),
-        make_temporary_output_clone(exec, output).get()));
-}
-
-
-template <typename ValueType>
-std::unique_ptr<Diagonal<ValueType>> MultiVector<ValueType>::extract_diagonal()
-    const
-{
-    const auto diag_size = std::min(this->get_size()[0], this->get_size()[1]);
-    auto diag = Diagonal<ValueType>::create(this->get_executor(), diag_size);
-    this->extract_diagonal(diag);
-    return diag;
-}
-
-
-template <typename ValueType>
 void MultiVector<ValueType>::compute_absolute_inplace()
 {
     this->get_executor()->run(
@@ -1747,21 +1681,6 @@ auto MultiVector<ValueType>::get_const_device_view() const -> const_device_view
 {
     return const_device_view{this->get_size(), this->get_stride(),
                              this->get_const_values()};
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::add_scaled_identity_impl(const LinOp* a,
-                                                      const LinOp* b)
-{
-    precision_dispatch_real_complex<ValueType>(
-        [this](auto dense_alpha, auto dense_beta, auto dense_x) {
-            this->get_executor()->run(multivector::make_add_scaled_identity(
-                dense_alpha->get_const_device_view(),
-                dense_beta->get_const_device_view(),
-                dense_x->get_device_view()));
-        },
-        a, b, this);
 }
